requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.18
djangorestframework==3.15.2
greenlet==3.1.1
idna==3.10
playwright==1.49.1
pyee==12.0.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
